move common schema
