+++ starknet_py/net/schemas/common.py
@@ -1,9 +1,10 @@
 from typing import Any, Mapping, Optional, Union
 
-from marshmallow import ValidationError, fields
+from marshmallow import Schema, ValidationError, fields, post_load
 
 from starknet_py.net.client_models import (
     BlockStatus,
+    StorageEntry,
     TransactionStatus,
     TransactionType,
 )
@@ -114,3 +115,13 @@ class TransactionTypeField(fields.Field):
             )
 
         return TransactionType(value)
+
+
+class StorageEntrySchema(Schema):
+    key = Felt(data_key="key", required=True)
+    value = Felt(data_key="value", required=True)
+
+    @post_load
+    def make_dataclass(self, data, **kwargs):
+        # pylint: disable=no-self-use
+        return StorageEntry(**data)
